Refactor endpoint into auth endpoint
