+++ backend/features/auth/routes/__init__.py
@@ -1,0 +1,1 @@
+from .auth_routes import router as auth_router
