+++ backend/routers.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter, FastAPI
 
+from backend.features.auth.routes import auth_router
 from backend.features.user.routes import user_router
 from backend.routes import items_router
 
@@ -9,5 +10,6 @@ def register_routers(app: FastAPI):
 
     api_router.include_router(items_router)
     api_router.include_router(user_router)
+    api_router.include_router(auth_router)
 
     app.include_router(api_router)
